requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
boto3==1.35.0
python-dotenv==1.0.0
pydantic==2.9.0